therapeutic_rag/requirements.txt
langchain==0.3.20
langchain_chroma==0.2.2
langchain_community==0.3.19
langchain_openai==0.3.8
python-dotenv==1.0.1
streamlit==1.43.2
